Add all contract implementation to all classes; add all MainPresenterTest cases.
I first started to add the test cases to MainPresenterTest and quickly found out that I needed to implement
the contracts (interfaces) for all of the Presenters and Object classes.

In doing so, I have delcared public overridden constructors for each class. The presenter classes each contain
one object of the object it is suppose to present for.

By this commit, the MainPresenterTest class is complete.