Remove list directory and replaced with checkbox_list directory.
I over complicated the process of storing the checkbox data. I have deleted all the list and list item files as they are now redundant.

I plan to use the CheckBoxHandler to store the List<CheckBox> in which I will use to store the 5 checkboxes on the main activity screen.